GUI Update 2
+ Buttons for play, deck editor, shop, settings, and quit to the main
menu. Only shop and quit have actual functionality as of now.
+ Created a general method for applying a drop shadow to an element in
the interface.
+ Created a new method called prepSubMenus as a possible way to manage
scene switching by preparing scenes that will be loaded when a button is
pressed. Might remove this on the next update in favor of a better
system.
+ Minor revision to main, having the stage be handed off to the
gameGraphics class to allow for scene switching functionality to be
added to the class.
+ Gave the player a shit-ton of currency by default to better test
card-pack RNG when they get added.
+ A bunch of graphical assets for later use.