General Game Update 2
+ Moved card image loading from the deck editor over to a method that is
called once during game initialization to improve overall performance
and memory consumption.
+ Moved deck editor sub-menu creation to its own method and added some
new visual effects.
+ Minor optimizations to memory consumption for both the deck editor and
shop sub-menus.
*** Significant bug discovered with the animated video background on the
main menu: memory is NOT freed at the end of a the video loop and
instead another video is created. This is resulting in increasingly high
memory consumption and CPU usage from the program and will need to be
patched before the video can be allowed to loop again.