Performance Update 2
+ What may have been previously thought to be a memory leak has now been
identified as just the JVM being very liberal in its memory allocation.
To reduce memory stress when testing include -Xmx512m as a JVM argument
to limit the JVM to 0.5 GB of RAM.
+ RAM usage tool has been added and runs every second, printing to the
console the memory currently being used by the program.
+ Deck editor should now function as intended, updating player card pool
when cards are purchased in the store.
+ Minor text fixes and more diagnostic output to the console during
runtime.