Arena UI Update 1
+ Added a base UI to build upon for custom arena game mode.
+ Began work on a foundation for creating custom game modes.
+ Custom game button will run a simple simulation of a game between two
bots when pressed and load in WIP assets for the arena.
+ Added a health bar for player objects.
+ New graphics assets.
* NOTE: Before trying to run this build, please add aiData.txt to the
main NetBeans project folder (the one that contains: build, dist,
nbproject, etc.) or the custom game will fail to launch. This is due to
me not loading this data file in the same way as monsterData.txt is
loaded and will be fixed in a later build to grab the file from internal
source folders.
* NOTE: To play the custom game mode without having a
NullPointerException thrown, please build a custom deck of cards in the
first slot with the deck editor before trying to play a game. Choosing
which custom deck to play with has not been added to the play menu yet.