General Game Update 4 (0.150)
+ 29 new cards have been added, this includes new cards for tiers 1->5
and the addition of spell cards. Card drops have been adjusted
accordingly to account for these new cards.

+ Work into creating multiplayer games has been started with the
creation of a new GameClient and Network class. The client currently
attempts to connect to the MBA server (which needs to be running on
localhost to work), and it also supports joining and leaving the
matchmaking queue. A non-functional multiplayer button has now been
added to the main menu to support these future games.

+ Game has been updated to version 0.150

* With the addition of a new multiplayer framework for this update,
compiling monster battle arena source now requires the addition of an
external library called Kryonet. This can be downloaded from the
libraries folder in the repository and then needs to be added to the
project in NetBeans by right-clicking on the project in the project
explorer, going to "Properties", then on the pop-up window, going to
"Libraries" and selecting "Add JAR/Folder" then adding the
kryonet-2.21-all.jar file.